Consolidate M2VAE implementation
